ref: Recode the Graph Structure

+++ src/dissector/tmp.c
@@ -1,6 +1,6 @@
 #include "dissector.h"
 
-int dissector_generic(struct packet_info *pi, const u_char *buffer, int len)
+int dissector_generic(struct packet_info *pi, const u_char *buffer, size_t len)
 {
 	(void)pi;
 	(void)buffer;
@@ -9,7 +9,7 @@ int dissector_generic(struct packet_info *pi, const u_char *buffer, int len)
 };
 
 __attribute__((weak)) int dissector_ethernet(struct packet_info *pi,
-					     const u_char *buffer, int len)
+					     const u_char *buffer, size_t len)
 {
 	(void)pi;
 	(void)buffer;
@@ -18,7 +18,7 @@ __attribute__((weak)) int dissector_ethernet(struct packet_info *pi,
 };
 
 __attribute__((weak)) int dissector_linux_sll(struct packet_info *pi,
-					      const u_char *buffer, int len)
+					      const u_char *buffer, size_t len)
 {
 	(void)pi;
 	(void)buffer;
@@ -27,7 +27,7 @@ __attribute__((weak)) int dissector_linux_sll(struct packet_info *pi,
 };
 
 __attribute__((weak)) int dissector_ipv4(struct packet_info *pi,
-					 const u_char *buffer, int len)
+					 const u_char *buffer, size_t len)
 {
 	(void)pi;
 	(void)buffer;
@@ -36,7 +36,7 @@ __attribute__((weak)) int dissector_ipv4(struct packet_info *pi,
 };
 
 __attribute__((weak)) int dissector_ipv6(struct packet_info *pi,
-					 const u_char *buffer, int len)
+					 const u_char *buffer, size_t len)
 {
 	(void)pi;
 	(void)buffer;
@@ -45,7 +45,7 @@ __attribute__((weak)) int dissector_ipv6(struct packet_info *pi,
 };
 
 __attribute__((weak)) int dissector_arp(struct packet_info *pi,
-					const u_char *buffer, int len)
+					const u_char *buffer, size_t len)
 {
 	(void)pi;
 	(void)buffer;
@@ -54,7 +54,7 @@ __attribute__((weak)) int dissector_arp(struct packet_info *pi,
 };
 
 __attribute__((weak)) int dissector_tcp(struct packet_info *pi,
-					const u_char *buffer, int len)
+					const u_char *buffer, size_t len)
 {
 	(void)pi;
 	(void)buffer;
@@ -63,7 +63,7 @@ __attribute__((weak)) int dissector_tcp(struct packet_info *pi,
 };
 
 __attribute__((weak)) int dissector_udp(struct packet_info *pi,
-					const u_char *buffer, int len)
+					const u_char *buffer, size_t len)
 {
 	(void)pi;
 	(void)buffer;
@@ -72,7 +72,7 @@ __attribute__((weak)) int dissector_udp(struct packet_info *pi,
 };
 
 __attribute__((weak)) int dissector_icmp(struct packet_info *pi,
-					 const u_char *buffer, int len)
+					 const u_char *buffer, size_t len)
 {
 	(void)pi;
 	(void)buffer;
@@ -81,7 +81,7 @@ __attribute__((weak)) int dissector_icmp(struct packet_info *pi,
 };
 
 __attribute__((weak)) int dissecotr_icmp6(struct packet_info *pi,
-					  const u_char *buffer, int len)
+					  const u_char *buffer, size_t len)
 {
 	(void)pi;
 	(void)buffer;
@@ -90,7 +90,7 @@ __attribute__((weak)) int dissecotr_icmp6(struct packet_info *pi,
 };
 
 __attribute__((weak)) int dissector_dhcp(struct packet_info *pi,
-					 const u_char *buffer, int len)
+					 const u_char *buffer, size_t len)
 {
 	(void)pi;
 	(void)buffer;
@@ -99,7 +99,7 @@ __attribute__((weak)) int dissector_dhcp(struct packet_info *pi,
 };
 
 __attribute__((weak)) int dissector_dns(struct packet_info *pi,
-					const u_char *buffer, int len)
+					const u_char *buffer, size_t len)
 {
 	(void)pi;
 	(void)buffer;
@@ -108,7 +108,7 @@ __attribute__((weak)) int dissector_dns(struct packet_info *pi,
 };
 
 __attribute__((weak)) int dissector_smtp(struct packet_info *pi,
-					 const u_char *buffer, int len)
+					 const u_char *buffer, size_t len)
 {
 	(void)pi;
 	(void)buffer;
@@ -117,7 +117,7 @@ __attribute__((weak)) int dissector_smtp(struct packet_info *pi,
 };
 
 __attribute__((weak)) int dissector_pop(struct packet_info *pi,
-					const u_char *buffer, int len)
+					const u_char *buffer, size_t len)
 {
 	(void)pi;
 	(void)buffer;
@@ -126,7 +126,7 @@ __attribute__((weak)) int dissector_pop(struct packet_info *pi,
 };
 
 __attribute__((weak)) int dissector_imap(struct packet_info *pi,
-					 const u_char *buffer, int len)
+					 const u_char *buffer, size_t len)
 {
 	(void)pi;
 	(void)buffer;
@@ -135,7 +135,7 @@ __attribute__((weak)) int dissector_imap(struct packet_info *pi,
 };
 
 __attribute__((weak)) int dissector_http(struct packet_info *pi,
-					 const u_char *buffer, int len)
+					 const u_char *buffer, size_t len)
 {
 	(void)pi;
 	(void)buffer;
@@ -144,7 +144,7 @@ __attribute__((weak)) int dissector_http(struct packet_info *pi,
 };
 
 __attribute__((weak)) int dissector_ftp(struct packet_info *pi,
-					const u_char *buffer, int len)
+					const u_char *buffer, size_t len)
 {
 	(void)pi;
 	(void)buffer;
@@ -153,7 +153,7 @@ __attribute__((weak)) int dissector_ftp(struct packet_info *pi,
 };
 
 __attribute__((weak)) int dissector_https(struct packet_info *pi,
-					  const u_char *buffer, int len)
+					  const u_char *buffer, size_t len)
 {
 	(void)pi;
 	(void)buffer;
@@ -162,7 +162,7 @@ __attribute__((weak)) int dissector_https(struct packet_info *pi,
 };
 
 __attribute__((weak)) int dissector_ssh(struct packet_info *pi,
-					const u_char *buffer, int len)
+					const u_char *buffer, size_t len)
 {
 	(void)pi;
 	(void)buffer;
@@ -171,7 +171,7 @@ __attribute__((weak)) int dissector_ssh(struct packet_info *pi,
 };
 
 __attribute__((weak)) int dissector_telnet(struct packet_info *pi,
-					   const u_char *buffer, int len)
+					   const u_char *buffer, size_t len)
 {
 	(void)pi;
 	(void)buffer;
@@ -180,7 +180,7 @@ __attribute__((weak)) int dissector_telnet(struct packet_info *pi,
 };
 
 __attribute__((weak)) int dissector_sctp(struct packet_info *pi,
-					 const u_char *buffer, int len)
+					 const u_char *buffer, size_t len)
 {
 	(void)pi;
 	(void)buffer;
@@ -189,7 +189,7 @@ __attribute__((weak)) int dissector_sctp(struct packet_info *pi,
 };
 
 __attribute__((weak)) int dissector_ldap(struct packet_info *pi,
-					 const u_char *buffer, int len)
+					 const u_char *buffer, size_t len)
 {
 	(void)pi;
 	(void)buffer;
